auth-service/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
httpx==0.28.1
PyJWT==2.8.0
python-dotenv==1.2.1
pydantic==2.12.4
